discogs_credits: extract review table
Issue #32 — biggest move yet. `showReviewTable` (~700 lines) moves to
`src/review-table.js`. The post-preflight UI where each Discogs credit
gets a row, auto-matches are pre-filled, attention-needed ones get
search + create actions. Renders into the UI bar's `<ul.logs>` and waits
for the user to click "Start import" before kicking off the dispatch.

Self-contained: imports `db`/`readIdbRecord` (storage), `mbThrottle`
(api-mb), `getDiscogsLinkKey`/`link_infos` (api-discogs),
`guessSortName` (mappers), `_hideBar` (progress-bar), `getLogContainer`
(log) — all already-extracted modules.

`logs.appendChild(…)` / `logs.scrollIntoView(…)` rewired to call
`getLogContainer()` (log.js's accessor) instead of relying on the
entry's `let logs` module-local.

`_urlCheckSessionCache` moved INTO the review-table module as a private
const — only ever read by showReviewTable, so it belongs alongside.

~720 lines moved. Smoke (Spiritual Jazz: 95 staged rels) green.
